requirements.txt
pydantic
matplotlib
openai
python-dotenv
fastapi
uvicorn